requirements.txt
pandas
python-dotenv
elasticsearch
